+++ source/index.js
@@ -1,3 +1,4 @@
+// @ts-check
 /* eslint no-use-before-define:0 */
 'use strict'
 
@@ -7,7 +8,13 @@ const textExtensions = require('textextensions')
 const binaryExtensions = require('binaryextensions')
 
 /**
- * @typedef {'utf8'|'binary'} EncodingResult
+ * WIll be `null` if `buffer` was not provided. Otherwise will be either `'utf8'` or `'binary'`.
+ * @typedef {'utf8'|'binary'|null} EncodingResult
+ */
+
+/**
+ * WIll be `null` if neither `filename` nor `buffer` were provided. Otherwise will be a boolean value with the detection result.
+ * @typedef {boolean|null} TextOrBinaryResult
  */
 
 /**
@@ -19,13 +26,13 @@ const binaryExtensions = require('binaryextensions')
 /**
  * @callback IsTextCallback
  * @param {Error?} error
- * @param {boolean} [result]
+ * @param {TextOrBinaryResult} [isTextResult]
  */
 
 /**
  * @callback IsBinaryCallback
  * @param {Error?} error
- * @param {boolean} [result]
+ * @param {TextOrBinaryResult} [isBinaryResult]
  */
 
 /**
@@ -39,14 +46,12 @@ const binaryExtensions = require('binaryextensions')
  * Determined by extension checks first (if filename is available), otherwise if unknown extension or no filename, will perform a slower buffer encoding detection.
  * This order is done, as extension checks are quicker, and also because encoding checks cannot guarantee accuracy for chars between utf8 and utf16.
  * The extension checks are performed using the resources https://github.com/bevry/textextensions and https://github.com/bevry/binaryextensions
+ * In a later major release, this function will become {@link isText} so you should use that instead.
  * @param {string} [filename] The filename for the file/buffer if available
  * @param {Buffer} [buffer] The buffer for the file if available
- * @returns {boolean}
+ * @returns {TextOrBinaryResult}
  */
 function isTextSync(filename, buffer) {
-	// Prepare
-	let isText = null
-
 	// Test extensions
 	if (filename) {
 		// Extract filename
@@ -57,84 +62,140 @@ function isTextSync(filename, buffer) {
 
 		// Cycle extensions
 		for (const extension of parts) {
-			if (textExtensions.indexOf(extension) !== -1) {
-				isText = true
-				break
+			if (textExtensions.includes(extension)) {
+				return true
 			}
-			if (binaryExtensions.indexOf(extension) !== -1) {
-				isText = false
-				break
+			if (binaryExtensions.includes(extension)) {
+				return false
 			}
 		}
 	}
 
 	// Fallback to encoding if extension check was not enough
-	if (buffer && isText === null) {
-		isText = getEncodingSync(buffer) === 'utf8'
+	if (buffer) {
+		return getEncodingSync(buffer) === 'utf8'
 	}
 
-	// Return our result
-	return isText
+	// No buffer was provided
+	return null
 }
 
 /**
- * Determine if the filename and/or buffer is text.
- * Uses {@link isTextSync} behind the scenes.
- * @param {string?} filename Forwarded to `isTextSync`
- * @param {Buffer?} buffer Forwarded to `isTextSync`
- * @param {IsTextCallback} next
+ * Callback wrapper for {@link isTextSync}.
+ * @param {string?} filename
+ * @param {Buffer?} buffer
+ * @param {IsTextCallback} callback
  * @returns {void}
  */
-function isText(filename, buffer, next) {
+function isTextCallback(filename, buffer, callback) {
 	let result
 	try {
 		result = isTextSync(filename, buffer)
 	} catch (err) {
-		next(err)
+		callback(err)
 	}
-	next(null, result)
+	callback(null, result)
 }
 
 /**
- * Determine if the filename and/or buffer is binary.
- * Uses {@link isTextSync} behind the scenes.
- * @param {string} [filename] Forwarded to `isTextSync`
- * @param {Buffer} [buffer] Forwarded to `isTextSync`
- * @returns {boolean}
+ * Promise wrapper for {@link isTextSync}.
+ * @param {string?} filename
+ * @param {Buffer?} buffer
+ * @returns {Promise<TextOrBinaryResult>}
+ */
+function isTextPromise(filename, buffer) {
+	try {
+		return Promise.resolve(isTextSync(filename, buffer))
+	} catch (err) {
+		return Promise.reject(err)
+	}
+}
+
+/**
+ * Wrapper around {@link isTextSync} for sync signature and {@link isTextCallback} async signature.
+ * In a later major release, {@link isTextSync}.will become this function, so if you prefer the callback interface you should use {@link isTextCallback}.
+ * @param {string?} filename
+ * @param {Buffer?} buffer
+ * @param {IsTextCallback} [callback] If provided, void will be returned, as the result will provided to the callback.
+ * @returns {TextOrBinaryResult|void} If no callback was provided, then the result is returned.
+ */
+function isText(filename, buffer, callback) {
+	if (callback) {
+		return isTextCallback(filename, buffer, callback)
+	} else return isTextSync(filename, buffer)
+}
+
+/**
+ * Inverse wrapper for {@link isTextSync}.
+ * In a later major release, this function will become {@link isBinary} so you should use that instead.
+ * @param {string} [filename]
+ * @param {Buffer} [buffer]
+ * @returns {TextOrBinaryResult}
  */
 function isBinarySync(filename, buffer) {
-	// Handle
-	const result = isTextSync(filename, buffer)
-	return !result
+	const text = isTextSync(filename, buffer)
+	if (text == null) return null
+	return !text
 }
 
 /**
- * Determine if the filename and/or buffer is binary.
- * Uses {@link isTextSync} behind the scenes.
- * @param {string?} filename Forwarded to `isText`
- * @param {Buffer?} buffer Forwarded to `isText`
- * @param {IsBinaryCallback} next
+ * Callback wrapper for {@link isBinarySync}.
+ * @param {string?} filename
+ * @param {Buffer?} buffer
+ * @param {IsTextCallback} callback
  * @returns {void}
  */
-function isBinary(filename, buffer, next) {
+function isBinaryCallback(filename, buffer, callback) {
 	let result
 	try {
-		result = isTextSync(filename, buffer)
+		result = isBinarySync(filename, buffer)
 	} catch (err) {
-		next(err)
+		callback(err)
 	}
-	next(null, !result)
+	callback(null, result)
+}
+
+/**
+ * Promise wrapper for {@link isBinarySync}.
+ * @param {string?} filename
+ * @param {Buffer?} buffer
+ * @returns {Promise<TextOrBinaryResult>}
+ */
+function isBinaryPromise(filename, buffer) {
+	try {
+		return Promise.resolve(isBinarySync(filename, buffer))
+	} catch (err) {
+		return Promise.reject(err)
+	}
+}
+
+/**
+ * Wrapper around {@link isBinarySync} for sync signature and {@link isBinaryCallback} async signature.
+ * In a later major release, {@link isBinarySync}.will become this function, so if you prefer the callback interface you should use {@link isBinaryCallback}.
+ * @param {string?} filename
+ * @param {Buffer?} buffer
+ * @param {IsTextCallback} [callback] If provided, void will be returned, as the result will provided to the callback.
+ * @returns {TextOrBinaryResult|void} If no callback was provided, then the result is returned.
+ */
+function isBinary(filename, buffer, callback) {
+	if (callback) {
+		return isBinaryCallback(filename, buffer, callback)
+	} else return isBinarySync(filename, buffer)
 }
 
 /**
  * Get the encoding of a buffer.
- * We fetch a bunch chars from the start, middle and end of the buffer.
- * We check all three, as doing only start was not enough, and doing only middle was not enough, so better safe than sorry.
+ * Checks the start, middle, and end of the buffer for characters that are unrecognized within UTF8 encoding.
+ * History has shown that inspection at all three locations is necessary.
+ * In a later major release, this function will become {@link getEncoding} so you should use that instead.
  * @param {Buffer} buffer
  * @param {EncodingOpts} [opts]
  * @returns {EncodingResult}
  */
 function getEncodingSync(buffer, opts) {
+	// Check
+	if (!buffer) return null
+
 	// Prepare
 	const textEncoding = 'utf8'
 	const binaryEncoding = 'binary'
@@ -182,28 +243,64 @@ function getEncodingSync(buffer, opts) {
 /**
  * Get the encoding of a buffer.
  * Uses {@link getEncodingSync} behind the scenes.
- * @param {Buffer} buffer Forwarded to `getEncodingSync`
- * @param {EncodingOpts} opts Forwarded to `getEncodingSync`
- * @param {GetEncodingCallback} next
+ * @param {Buffer} buffer
+ * @param {EncodingOpts} [opts]
+ * @param {GetEncodingCallback} callback
  * @returns {void}
  */
-function getEncoding(buffer, opts, next) {
+function getEncodingCallback(buffer, opts, callback) {
+	if (typeof opts === 'function' && callback == null)
+		return getEncodingCallback(buffer, null, opts)
 	/** @type {EncodingResult?} */
 	let result
 	try {
 		result = getEncodingSync(buffer, opts)
 	} catch (err) {
-		next(err)
+		callback(err)
 	}
-	next(null, result)
+	callback(null, result)
+}
+
+/**
+ * Promise wrapper for {@link getEncodingSync}.
+ * @param {Buffer} buffer
+ * @param {EncodingOpts} [opts]
+ * @returns {Promise<EncodingResult>}
+ */
+function getEncodingPromise(buffer, opts) {
+	try {
+		return Promise.resolve(getEncodingSync(buffer, opts))
+	} catch (err) {
+		return Promise.reject(err)
+	}
+}
+
+/**
+ * Wrapper around {@link getEncodingSync} for sync signature and {@link getEncodingCallback} async signature.
+ * In a later major release, {@link getEncodingSync}.will become this function, so if you prefer the callback interface you should use {@link getEncodingCallback}.
+ * @param {Buffer} buffer
+ * @param {EncodingOpts} [opts]
+ * @param {GetEncodingCallback} [callback] If provided, void will be returned, as the result will provided to the callback.
+ * @returns {EncodingResult|void} If no callback was provided, then the result is returned.
+ */
+function getEncoding(buffer, opts, callback) {
+	if (callback || typeof opts === 'function') {
+		return getEncodingCallback(buffer, opts, callback)
+	} else return getEncodingSync(buffer, opts)
 }
 
 // Export
 module.exports = {
 	isTextSync,
+	isTextCallback,
+	isTextPromise,
 	isText,
 	isBinarySync,
+	isBinaryCallback,
+	isBinaryPromise,
 	isBinary,
+	getEncoding,
 	getEncodingSync,
-	getEncoding
+	getEncodingPromise,
+	getEncodingCallback
 }